remove deprecated PortBindingPort mixin
the PortBindingMixin is a deprecated mixin for legacy out-of-tree
neutron plugins and not used by networking-aci driver since it's never
used. But it's still installs query hooks degrading the port-fetch
performance and breaking host-search.

See also commit #7422903 in neutron:

This patch moves PortBindingPort model under neutron/db/models for
consistency.There are no plans to implement an OVO object for the model
at this point in time. Actually, the model should probably eventually go
away from Neutron tree since it's not used by reference plugin, and is
just a remnant from the times when the neutron tree contained lots of
plugins, some of which were using the model.