Refactor and fix method to find clearable physdoms
_get_clearable_phys_doms() had various issues after it being refactored:

Wrong binding level: We need to search for binding level 1 (as stated in
the comment), not 0. Additionally, we don't need the level being passed
to this method, therefore we remove the argument.

Wrong dict key physdoms being used with other_binding_host_config, it
needs to be 'physical_domain'.

Modifying a list while we're iterating over it: This part has been
refactored to use sets, which also leads to the list being copied and
this problem not arising anymore.