Prevent excess instanciation of DBPlugins
Every call of get_address_scope_name() instanciates a DBPlugin(). After
the function is done the DBPlugin is not deleted, as it is still
referenced somewhere. This leads to excess instances that slow down
neutron-server significantly. With this patch DBPlugin() will be
instanciated a single time and passed on to where it is needed.