Get alternate binding host from binding profile on port delete
When the driver checks if a physdom can be deleted it checks if there
are any ports left on the segment. For this the binding host of the port
is relevant. In the case of baremetal ports this binding host is
obtained from the binding profile (as it is done in other parts of the
driver, see usage of switch_from_local_link()). Therefore we need to
consider this alternate datasource as well when checking if a physdom is
still in use.