Drop allocations_manager, use neutron-internal ml2_vlan allocations code
This commit drop the allocations_manager in favor of the ml2 VLAN type
driver that does the same thing inside neutron, it creates dynamic
network segments.

This commit also allows leverage of the network_segment_range plugin of
neutron. It also fixes racecondition where multple segments been
allocated for the same physical networks as well as the releasing of
vlan segments if not required anymore.