Referential integrity for AllocationsModel
To prevent orphans in the db that block vlan ids we introduce
foreignkeys. When a network is deleted network_id gets nulled, when a
networksegment is deleted segment_id gets nulled. To prevent cases where
a networksegment was deleted but the network of the segment still exists
we now regard an allocation as free when it has a nulled segment_id.