Reimplement and make consistent use of retry logic
A retry needs to be done when the APIC does not want to talk to us
anymore. This might be caused on an internal error, a locked config
object, a broken APIC (and we need to switch over) or due to an auth
token timeout.

Before this commit retrying was only used inconsistently for certain
queries, now it is used for all queries done by the driver.