Improve get_all_epgs() performance with better filtering
get_all_epgs() previously iterated over all tenants, then got all EPGs
of that tenant. As it got everything without subtree=full, this could be
done in ~8s in qa.

Now we just get all EPGs from the ACI that have a tenant starting with
the preconfigured tenant-prefix, which reduces the query time in this
instance to 200ms.