Renew APIC token before it times out
We don't need to wait until our token times out and a query failure
triggers a relogin. We can just look at our token timeout and see if it
is still valid. If it is bound to timeout in the next 120s
(configurable) then we call reauth() to refresh it. If it's invalid we
trigger a relogin, but this should only happen when the syncloop is
turned off or something else hangs.