Return None on not found for single-mode in cobra client
When called with single=True, query() unpacks the result. If the result
is empty then we will return None instead of an empty list.

For getting a BD/EPG we want to have at least a debug log message so we
throw a runtime error that we then catch and log.