Avoid race condition in token refresh with locks
When two threads concurrently try to refresh the auth token, then one
might fail. Along with it its query fails. As we don't want that we
protect the refresh with a lock. To avoid two refreshes we check inside
the lock if another thread did already refresh the token and do nothing
if it has.