Direct-on-aci portbindings with mode-switching API
Direct-mode is a new config option for a hostgroup. A hostgroup in
direct mode is expected to be directly bound by the driver. For these
hosts we still do an initial hierarchical portbinding but bind said port
in the second bind_port() call.

Direct-mode hostgroups can have two modes, which change how a port in
this port is being bound:

MODE_INFRA: A port in this group will be bound like a normal VM port,
except that on the ACI side the encapsulation mode will be access /
untagged instead of trunk / tagged. This mode can be used to install a
hypervisor host via ironic. A MODE_INFRA hostgroup will share a physical
domain and vlan-pool with a parent hostgroup. This parent hostgroup is
is used for regular VMs, is not a direct-mode hostgroup and will result
in normal tagged bindings (as always). A portbinding of a port in
MODE_INFRA will take precedence, i.e. its bindings will be deployed as
untagged in an EPG if deployed on a network where a port in the parents
hostgroup is already deployed.

MODE_BAREMETAL: A port in this segment will be bound on its own segment,
have its own physical domain and vlan pool. A custom VPC policy group
will be created and all port selectors referenced in this hostgorup will
be assigned to this policy group. Ports will be bound untagged. Bindings
of a hostgroup in MODE_BAREMETAL will be removed from its parent group.

An API for the ACI driver is now present. It allows to get insight in
the current config (more of an API test endpoint, but might be useful in
the future).
This API also contains means to list and set the hostgroup mode of all
direct-mode hostgroups. On switching the ACI agent will be notified to
perform changes on the ACI, switching interface profiles etc..

To better support these new modes the rpc api support call
_get_network() was refactored. It now gets all hosts on a network and
adds hostgroup-based config accordingly to its response. This should
result in a faster call and less code to maintain.