Add trunk support for MODE_BAREMETAL
The user can now create trunks and subports for baremetal hosts
(meaning hosts in an hostgroup that is in MODE_BAREMETAL). The user can
attach OpenStack networks with vlan ids of their choosing as trunk ports
to a server. The separate physical domain in ACI already used by
MODE_BAREMETAL allows us to use a completely separate vlan pool.