Assume direct_mode false for hostgroups in agent
Fixed bindings don't provide a direct-mode (it is not applicable for
them). Instead of adding a direct_mode entry in their config dict we
just default to False in the agent for direct_mode. This has also the
nice side effect that on the transition from non-direct_mode aware
neutron-server instances to direct_mode-aware we don't get any
exceptions regarding missing keys.