Replace hash-ring with uhashring
The hash-ring pypi package is not python3 compatible and looks like it
is out of maintenance for quite some time. uhashring on the other hand
is under active development and python3 compatible.

We need to make sure that by replacing the implementation we do not
generate different hash values, as this would cause a massive
reallocation of networks. Therefore I added some tests with 100 values
of the original implementation that we can compare against.

uhashring generates the same results as hash-ring when using the Ketama
hash function with replicas=3 (and implicitly vnodes=40). Specifying
replicas is available since uhashring v1.2.

This commit breaks python2 compability, as uhashring 1.2 is only
available for python3.