Get tox tests running without any extras
We need to remove aci dependencies from the requirements, as with them
present we cannot run the tests and they are not published by Cisco.

We also need to pin the pip version, as at the time of this writing URL
dependencies are not supported. When they will be supported by a current
release this can be removed again. See https://github.com/pypa/pip/issues/8253