Add acicobra as an extra dependency
We remove acicobra and acimodel from requirements.txt as they are not
necessary to run the tests. They are only necessary to run the agent
part of the driver.

Due to the way of how Cisco handles the distribution of their we cannot
directly install these packages as a dependency. Cisco does not publish
them anywhere publicly. These packages need to be gathered from an APIC
via http or put on an internal pypi repo just for installing this
driver.

Therefore, to now install the driver with the aci* packages, please
install networking_aci[acicobra].