Handle excetion without message on connection error
When a connection error to the APIC occurs we rotate to the next
controller. Now we are in a position where our collection of fallback
exceptions does not have a message attribute, causing this failover
to... fail. Hence we now check for a message attribute and if none is
present just return the string representation of the exception.