Fix off-by-one in segment/vlan range calculation
Ranges are converted from ther a:b,c:d,... string format in the config to a
form of [(a, b), (c, d), ...] format to keep in the config. They are
later converted to a set containing all numbers in these ranges. Both
conversion methods accomodated for python's range() being
right-exclusive, so they both added +1 on the right side, which leads to
the off-by-one. We now remove the +1 from the initial conversion.

The intermediate format used here is present because we don't want to
keep the set of all numbers around, as it makes the data structure much
bigger, which is annoying when transporting it via RPC.