Use valid version for acicobra/acimodel
4.2-2.68 is not a valid version specification for recent versions of
pip. So, besides the acicobra/acimodels package not being publicly
available it also has an invalid version number, meaning we cannot just
install it as a dependency without repackaging it. As the minor version
of the package does not matter for us at the moment we'll just require
it to be >=4.2.