Scope direct-mode BM-on-ACI to project
Due to a variety of ACI bugs, limitations, problems and general
unwillingness we can't use the same VLAN from different pools in the
same EPG. This breaks a lot of assumptions the original direct-mode
BM-on-ACI code was based on. But to solve this we are going to...

Instead of creating BM-entities on a per-node basis, all of them are now
created on a per-project basis. We select the project the first
portbinding for a node is made to. The entities are named
`$prefix-$project_id`. This also means we need to redo the lifecycle
management of certain objects:

Networksegments: A networksegment is now created for each
network-project combination with an ACI BM port. When no port remains on
this network-project segment the segment can be deleted. Baremetal
segment allocation has been completely rewritten and an abstract of how
it works now can be found in the docstring of the method.

Binding/VPC infra/BM mode: The mode cannot be switched anymore on
external-API mode switch, as we need to reset it whenever an OpenStack
BM port changes projects (i.e. redeploy of instance). Therefore the
settings are only applied to the VPC when the first port with a host
matching the BM hostgroup is deployed and reset to infra-mode whenever
the last port for this hostgroup is removed. These checks heavily rely
on checking all bindings done to segments for the `$prefix-$project_id`
physical network (aka this is how we find all ports for a node).

Physdom-EPG association: A physdom will be associated to an EPG as long
as there is a port left for it to be mapped.

To implement the new ACI restrictions there are a lot of checks done
both on trunk subport addition and on baremetal segment allocation:
 * projects now have to provide vlan consistency
 * networks cannot have mixed VLANs from same pool anymore

This means if the user chose vlan 1000 for a network, subsequent
portbindings will have to use vlan 1000 for this network as well and
vlan 1000 cannot be used in another project. Also if one network has an
access port no trunk port is allowed in this network and vice versa. If
required, this is something we could - in part (maybe) - work around,
but it was not mentioned as a hard requirement.