Fix field query order for sa 1.2.1 compat
With Neutron Ussuri / SqlAlchemy 1.3.1+ the old original field order for
this query worked as intended. With 1.2.1 SqlAlchemy takes the first
field queried as table for the FROM clause. This results in
networksegments being in the FROM clause, causing an SQL error
(1066, "Not unique table/alias: 'networksegments'"). It is very nice
that SA 1.3.1+ can handle this, though from a programming and
code-reading perspective it makes much more sense to pull the port
fields in front of the query. Therefore we'll be doing this patch both
for Ussuri and Queens.