Fix providing contract bug on external net create
When an external subnet is created in OpenStack the external network it
belongs to is created/updated and we make sure that it has the
appropriate providing/consuming contracts. Due to a refactor the
creation of providing and consuming contracts were reduced to only
create consuming contracts, which should now have been rectified. This
only causes problem with external networks that have been created after
deployment of this commit, but as there is no syncloop ensuring external
network config these networks need to be synced manually.

The bug was introduced in 5fd5931717e96efe9020d8fc0d948af327284984