Fix EPG cleanup for infra-mode bindings
When we remove an infra-mode binding (i.e. a direct mode hostgroup in
infra mode) we need to remove its bindings from the EPG even when the
segment is not released (as long as it is not in use on this segment
anymore). This needs to be done as infra-mode groups don't have their
own segment, but share their network segment with their parent
hostgroup. As an example node001-bb257 will be bound on the
networksegment bb256. When the node is removed, its bindings need to be
removed, while bb256 might stay due to other hosts being bound on that
segment.

As another note we do not support a binding of a host in the parent
hostgroup (nova-compute-bb257) and a subgroup (node001-bb257) in the
same network. It will work, but VMs on node001-b257 would not be
reachable in this particular network. Other hosts in that network should
be fine, so that's why this is not disallowed by default in the driver
code. Dragons may apply.