Honor AZ hints in portbindings
Add a (create|update)_port_precommit hook that checks if a port can be
bound to a certain hostgroup. Hostgroups need to be annotated with a
list of AZs they are in for this to work. A restriction of one AZ hint
per network is put on networks. AZ checks on portbind can be disabled
with the config option "az_checks_enabled". Hostgroup and network AZ
hints need to match up, so hostgroups with multiple AZs set cannot have
ports from networks with AZ hints bound in them, except if there is a
more specific AZ hint. Via az_hints an AZ per binding host can be set.

AZ checks are disabled for external networks, as these are allowed to be
stretched across AZs. One reason for this is to still have FIP mobility
and allow the customer to move the same FIP from their infra from one AZ
to another.