Finalize bindings for hostgroups without a 2nd level
This allows ports in a hostgroup to be directly bound on ACI without
having to deal with the ACI direct_mode. ACI direct mode is explicitly
thought out for server directly connected to ACI, but in some cases we
just need to make a vlan id available via a trunk, just like in normal
binding mode, but where no second level driver is available. "Finalize
bindings" allows us to do just that - finalize a binding directly on
ACI. These hostgroups will also not show up in the list-hostgroup-modes
API.