Fix PortBindingLevel join condition for multiple portbindings
We try to find out which binding hosts are on a network by asking the
DB, specifically the models PortBinding/PortBindingLevels. For migration
purposes a portbinding might have two portbindings at the same time,
meaning it has two entries in the PortBinding's table. Each entry has
its own "host", but in our get_hosts_on_network() method, we only did the
join between PortBinding/PortBindingLevel by port_id, not by port_id AND
host. This resulted  in the wrong binding/segment getting mapped to the
wrong host. Now we join on both fields and have no longer alternating results
for this join.

The whole bug resulted in some bindings being deleted and readded to an
EPG that had one of these ports in it, essentially flapping the binding
around and adding a wrong vlan into the EPG. This should now be fixed.