Add sync_network() call to agent RPC API
There are some scenarios where the Neutron server wants to trigger a
network sync on the ACI and now we can! sync_network() will exactly do
that by calling the same sync method the agent syncloop is using.

This can then be used to implement on-demand network syncs to react on
events or to implement a sync API into the driver.