Sync network on fabric driver transit creation
The fabric driver (networking-ccloud) is designed to allow communication
with the ACI fabric via transits. This means that the fabric driver
allocates transits + transit segments after a network create and then
signals other drivers that this has happened. The ACI driver needs to
react and make sure the EPG is in sync with that.

Transits on ACI side are defined as extra hostgroups with
"fabric_transit = true" set. These groups cannot be bound externally.
On network sync a the ACI driver normally puts all hostgroups into the
EPG that have at least one bound port to a segment in the network. As
transit segments do not have bound ports they need special handling. For
these we look for all segments with physnets that belong to a transit
hostgroup. If found, these segments are put into the network sync data
and therefore put into the EPG.

To react on the transit creation note by the fabric driver the ml2
driver's RPC interface needed to be made available to the driver itself
so we can call _get_network() and get the network sync data. As this is
only a simple class without any depenencies (e.g. baseclasses starting
obscure cleanup looks) we can just instanciate it in the driver's
__init__(), just after instanciating the db plugin (instead of only
instanciating it when running in rpc-server mode).

Augmenting get_hosts_on_network() in the db plugin is a little bit
clunky because of the no-bound-ports case. We essentially look for
segments with a physnet belonging to a transit group and then adds these
hostgroups by adding host and segment to the list returned by the
method. This is somewhat a back-and-forth that, but I chose to solve it
this way to keep changes in other places to a minimum (in this case in
_get_network()).