Fix "No mos in config request" CommitError in agent
In certain circumstances _get_network() could provide an entry that
contains a config without bindings in it. This can happen when all
bindings of a non-direct-mode hostgroup are used in the same network in
direct-mode. As we cannot mix tagged and untagged bindings inside an EPG
the untagged bindings take precedence. If we remove all bindings of a
hostgroup the agent does not want to commit this (why should it, there
is nothing to commit). ACI raises a config error here. To prevent that
we now only do a commit for the bindings once we actually have something
to commit.