Allow gateways to move off ACI
The ACI driver now has a config switch that indicates if it is
responsible for all l3 gateways. If turned off, we will not configure
any l3 gateways on ACI.

The other option for a network to not have its gateway on ACI is by it
being tagged with "gateway-host::cc-fabric". networking-ccloud will
throw an event whenever it wants a gateway to be moved from ACI to the
cc-fabric. We react on that event by tearing down all subnets from that
network. The network will still have some different properties set, as
it is still handled as an external network (but without any l3
gateways). We'll see if that is a good idea or if this is something
that'll need to change. Experience will show.