Disable dynamic segment filter
We have some segments that are created with the dynamic flag and some
that are not. In sapcc's forked neutron the dynamic segment filter is
disabled by default (set to None), but we can easily add this filter
in the ACI driver so we might one day drop our downstream patch.