Don't log an error on "invalid" binding profile
We used to log an error if the binding profile didn't contain a binding
host (as we have it for ports created by ironic) but nowadays the
binding profile is used for many more cases and doesn't always contain a
host. Also, we never needed to act on a "broken" binding profile. If we
would need to do that there are many other means to find it. Therefore
we lower the log level from error to debug.