Use context provided by request for client -> agent RPC
In our mechanism driver and in our API we used to use a pre-generated
context. This means that all API and RPC calls are scoped to this one
context and all log messages regarding these calls are scoped to the
request id of this one context. As this makes debugging far more
complicated, we're now using the provided context.