Use provided context for agent --> server RPC
Calls made from the agent to neutron server also used a one-time created
context instead of the context associated with the call, leading to one
single request id for all calls made by the agent, instead of them being
tied to a client request for which they were made.

Periodic functions create one context object per run, so we can identify
which calls belong to which run of the loop.