Resync network on direct port binding deletion
Originally we didn't anticipate that a infra-mode direct portbinding
of a node with the parent group bound in the network could be deleted,
thus we placed an error message in the driver, logging that this is
unsupported behavior. Now, as we got a pretty solid use-case (the parent
group provides infra to the to-be installed single node, like network
agents), we need to properly implement this. Luckily, this is something
we only need to fix on port deletion.

The issue with port deletion in this case is that we have a binding in
access mode, which needs to be reset to trunk mode (or "untagged" vs
"regular" in ACI speak). This could be done by calling bind_port() on
the parent group, but testing this implementation showed that if a user
is deleting multiple direct-mode bindings at the same time we might end
up with a race condition where the second call would undo what the first
call has done.

Therefore this is now implemented as a network sync via
sync_network_id(), where we only tell the agent which network should be
synced. The agent then grabs the current state from any
neutron-rpc-server instance, which should end up including all latest
changes.