Core 2.1: Manage AZ-local network prefix lists
For external networks we now need to manage a list of Match Rules (ACI
UI name) or in model-speak SubjP/MatchRtDest objects for each prefix of
a subnet with an address scope of an AZ aware network (network with AZ
hint).

This commit is still missing the overarching syncloop that cleans these
lists, as the lifecycle for the whole list is not based on a
network/EPG/BD anymore, as everything else is so far in the driver. For
syncing these lists we will need to fetch all subnets of an Address
Scope. This will most likely fall together with the lifecycle of another
object that will be introduced with the Core2.1 featureset.