Core 2.1: Manage Nullroutes
Nullroutes are routes for external subnets that have a subnetpool where
the subnet CIDR has no equal match in the subnetpool's list of prefixes.
The matching parent prefix in the subnetpool's list of prefixes will be
used as nullroute.

The nullroutes are set on all leaves where the subnet has a bound port
on. The lifecycle of the routes is not time critical (or at least that
is the current assumption I have been given to work with), therefore we
only handle routes in a syncloop, where we sync all nullroutes every 15
minutes.