Core 2.1: Syncloop for AZ-local network prefixes
Sync all AZ-local subnets with an address scope into their respective
list. This is the pendant to what we do on subnet creation or deletion
and cleans up old prefixes if something went wrong. An API is included
as well to run a sync or see the data in the API.

This also irons out some bugs with the subnet creation code (e.g. using
ip instead of prefix on the MatchRtDest).