Deduplicate CIDRs for nullroute sync
As one Subnetpool might have multiple subnets for a single prefix we
would have the same CIDR in the list multiple times. This would lead to
multiple API create requests for the same route. While this is not a
problem for the ACI API (i.e. we state the same intend multiple times)
it is also not... nice. Therefore we do proper deduplication of the
CIDRs now.