Reduce likelihood of syncloop race condition
When syncing networks we fetch a chunk of networks (e.g. 10 at a time).
These networks are then synced one by one (takes about 30 seconds in an
affected region). If in the meantime a network change was made, the sync
might roll back these changes, resulting in network outages in some
cases.

To reduce the likelihood of this we now introduce a "dirty networks"
cache, that keeps a list of all networks that have been altered since
syncloop start. If a network was altered, its data will be fetched
again. If it was altered again in this period in time it will be synced
regardless. A warning will be logged.

This is not an ideal solution, as it is not entirely thread safe, it
only reduces the likelihood, but in the complexity vs. impact
estimation, it seems like a good first step. To properly implement this,
one option would be to introduce per-network locking and then let either
the syncloop wait (or skip) or let the RPC call wait for the other to be
done.

Another not thread safe solution would be to move fetching the network
closer to the sync (reducing complexity by leaving out the dirty
networks cache), which on one hand would be a small performance impact
(about 80ms extra per network in a short non-representative test), but
would require some refactoring on the other hand.