Also configure nullroutes for FIPs
FIPs do not have to be in the same subnet as router_gateway interfaces.
Currently we are only creating a nullroute if we find at least a
router_gateway with an ip allocation in the subnet. Nevertheless FIPs
do not have to be in the same subnet as the router_gateway. In that case
we would not configure the nullroute.

This commit queries all FIPs that are in subnets that we deemed relevant
in step 1 and adds the agent's hosts to the list.