Do not configure hostgroup mode if port-selectors are missing
Port-selectors are only needed if we disassemble the port-selector
pc-policy relationship when we come and go from baremetal mode. If a
host is not planned to ever go into baremetal that relationship is
already formed and will never be destroyed, we can thus bypass that
code.

Also adding a return if `port_sel_entities` is empty so we do not throw
an empty list to the API.