Properly delete unwanted subnet on network sync
When syncing a network we are cleaning a BD as well. If the BD has extra
subnets we call ensure_subnet_deleted(). Somehow we missed to pass the
address scope name and the network az along to this call when we did the
core 2.1 refactoring, resulting in a TypeError here.