Fix login URL for ACI 5.x compatibility
The /api suffix for the login URL is not needed and breaks with ACI 5.x.
Luckily we can just omit it and it seems to be working again. As our
internal debug toolkit has also a "special" acicobra abstraction which
also doesn't include the /api suffix, we have already tested this
working in some of our production regions, which are still on an older
verison of ACI.