Fix and improve allocation cleaning sync
This is a rewrite of the sync allocations algorithm, which should be
about 80% faster (due to different way to query the DB), hopefully
better to understand and should actually clean allocations up (as before
the deletion happened outside of a transaction and was never actually
committed).

I've started out to understand the original sync algorithm and why it is
so slow and arrived with a rewrite of it in the way. The rewrite ended
up being 5% faster (which is basically nothing in this context), but
hopefully it is better to understand. While profiling my code it came to
light that about 80% of the algorithm was spent creating all the
AllocationsModel instances. Querying only the fields that we need and
by that not creating the model instances makes it much faster. As we
don't need the instances anyway with the rewrite (we're doing most of
the aggregation via set operations) changing this works quite well.

The code now actually also has tests that it does what it should do.

In the long run we want to remove the custom allocation management from
the driver, which also should be possible as segment allocation in
Neutron is less (not?) prone to race conditions anymore since Yoga.
Transitioning without interrupting the ability to create new segments
and without introducing segmentation inconsistencies is not trivial
though, which is why we go with the sync rewrite before developing the
whole migration.