Only execute allocation sync once per pod/system
We now make sure that the allocations (and hostgroup modes) are only
synced once per system - in our main use-case this is once per
neutron-server pod in Kubernetes. This is done by having the first
process to start the sync create a file after it is done. If that file
is found, subsequent neutron-server processes will not initiate the sync
on the same system. To make sure we don't end up in another race
condition with this we use a external_lock() from oslo.concurrency,
which guards the critical code section.