Rotate ACI on ReadTimeout exception
We had a couple of cases where the ACI API connection failed due to
ReadTimeout exceptions and reconnecting did not solve the problem. In
these cases it might be wise to switch to the next device in our
rotation. We do this by adding the ReadTimeout to our list of
FALLBACK_EXCEPTIONS.