Fix L3 BD subnet cleaning
When cleaning the L3 BD of an external subnet we need to make sure we
pass along the correct address scope (again). The last fix only resulted
in us using the address scope from the last iterated subnet in the code
block over our block, but in cases where we don't have any subnets there
(e.g. gateway was moved to another fabric) this would result in an
UnboundLocalError for neutron_subnet. To fix this, we now iterate over
all subnets and take the first not-None address scope to pass along.
This should work, as in our infrastructure all subnets of a network must
either have the same or no address scope.