Ignore EVPN networks for nullroutes/az aware routes
When selecting the external subnets which will be part of the
nullroute/az aware subnet routes sync we now ignore all subnets of
networks that are tagged with gateway-host::cc-fabric. These tagged
networks have their gateway on the EVPN side of the fabric.

The question of if we should install a nullroute for a subnetpool with
a prefix that contains an external network that has its gateway on the
EVPN side the answer was yes, yes we should do this. Therefore it should
be fine to just exclude the networks from the DB query.