Make sure internal networks get their VRF
For an internal network to work properly, it needs a VRF association,
meaning we need a fvRsCtx associated to the BD. Before this commit, this
was part of ensure_internal_network_configured(), which is called as a
create_subnet_postcommit() hook. If that gets lost for any reason or
maybe (I'm not sure of this) is faster than the original
ensure_domain_and_epg() call, it might be the case that the RsCtx gets
lost and the EPG is not usable with the following ACI fault:

Configuration failed for $epg_dn due to VRF Not Present,

To reduce the likelihood for internal networks, we now make sure to also
create the fvRsCtx as part of ensure_domain_and_epg(). For external
networks, we need subnets anyway and will keep it "the old way" of
creating it as part of the create_subnet_postcommit() hook. If creating
external subnets becomes a more regular operation and is indeed facing
the same problem, we can revisit this decision.