Recompute api token validity inside refresh lock
When we check the token validity and acquire the token refresh lock we
need to make sure we compute the validity again, as it might have
changed due to another thread refreshing the lock already. If we don't
do that we might end up in a re-login loop.

If we have many threads sending requests to the APIC and the token
validity falls under the refresh threshold, multiple threads might try
to refresh the token. If now the APIC is under heavy load and takes a
long time to answer the refresh requests we might end up in a situation
where our session actually times out, triggering a re-login. Now that we
have multiple threads waiting their turn with refreshing the token, they
all try to re-login, causing the refresh time to be None (newly created
mo_dir without a logged in session), which again will cause new threads
to issue a re-login. Recomputing the token validity after acquiring the
lock should prevent that.

To better monitor cases like this we now also report if the refresh time
is None.