Only warn when trying to tag non-existent network
When the agent tries to tag a network via RPC that does no longer exist,
we used to get a DBReferenceError like this:

Cannot add or update a child row: a foreign key constraint fails
(`neutron`.`tags`, CONSTRAINT `tags_ibfk_1` FOREIGN KEY
(`standard_attr_id`) REFERENCES `standardattributes` (`id`) ON DELETE
CASCADE)

This means the standardattributes belonging to the network have been
deleted (alongside with the network). This is something we can just
ignore, as we don't need to tag something that is already gone and the
agent will never ask for it again. Therefore we write out a warning (in
case someone ever DOES have a problem with this) and carry one as usual.